move displayurlprovider to new module, update github branch source
